Add new tests
