+++ analysis/test/resources/data_analysis/templates/template_issues_using_diff/filtered_submissions_python3_hyperstyle.csv
@@ -19,3 +19,13 @@ print('%.5f' % e)
 print('%.5f' % e)
 
 ",2,1,1,1,"{""quality"": {""code"": ""EXCELLENT"", ""text"": ""Code quality (beta): EXCELLENT""}, ""issues"": []}","{""quality"": {""code"": ""EXCELLENT"", ""text"": ""Code quality (beta): EXCELLENT""}, ""issues"": [{""code"": ""WPS446"", ""text"": ""Found approximate constant: 2.718281828459045"", ""line"": ""eul = 2.718281828459045"", ""line_number"": 1, ""column_number"": 6, ""category"": ""BEST_PRACTICES"", ""difficulty"": ""MEDIUM"", ""influence_on_penalty"": 0}, {""code"": ""SC200"", ""text"": ""Possibly misspelt word: `eul`"", ""line"": ""eul = 2.718281828459045"", ""line_number"": 1, ""column_number"": 1, ""category"": ""BEST_PRACTICES"", ""difficulty"": ""MEDIUM"", ""influence_on_penalty"": 0}]}","{""quality"": {""code"": ""EXCELLENT"", ""text"": ""Code quality (beta): EXCELLENT""}, ""issues"": [{""code"": ""WPS446"", ""text"": ""Found approximate constant: 2.718281828459045"", ""line"": ""eul = 2.718281828459045"", ""line_number"": 1, ""column_number"": 6, ""category"": ""BEST_PRACTICES"", ""difficulty"": ""MEDIUM"", ""influence_on_penalty"": 0}, {""code"": ""SC200"", ""text"": ""Possibly misspelt word: `eul`"", ""line"": ""eul = 2.718281828459045"", ""line_number"": 1, ""column_number"": 1, ""category"": ""BEST_PRACTICES"", ""difficulty"": ""MEDIUM"", ""influence_on_penalty"": 0}]}"
+5,3,2021-02-25 09:00:00.000000,python3,"x = 1
+# x = 2
+# x = 3
+print(x)
+",3,2,1,1,"{""quality"": {""code"": ""EXCELLENT"", ""text"": ""Code quality (beta): EXCELLENT""}, ""issues"": []}","{""quality"": {""code"": ""EXCELLENT"", ""text"": ""Code quality (beta): EXCELLENT""}, ""issues"": [{""code"": ""WPS432"", ""text"": ""Found magic number: 1"", ""line"": ""x = 1"", ""line_number"": 1, ""column_number"": 5, ""category"": ""BEST_PRACTICES"", ""difficulty"": ""MEDIUM"", ""influence_on_penalty"": 0}]}","{""quality"": {""code"": ""EXCELLENT"", ""text"": ""Code quality (beta): EXCELLENT""}, ""issues"": [{""code"": ""WPS432"", ""text"": ""Found magic number: 1"", ""line"": ""x = 1"", ""line_number"": 1, ""column_number"": 5, ""category"": ""BEST_PRACTICES"", ""difficulty"": ""MEDIUM"", ""influence_on_penalty"": 0}]}"
+6,3,2021-02-25 09:00:00.000000,python3,"# x = 1
+x = 2
+# x = 3
+print(x)
+",4,3,1,1,"{""quality"": {""code"": ""EXCELLENT"", ""text"": ""Code quality (beta): EXCELLENT""}, ""issues"": []}","{""quality"": {""code"": ""EXCELLENT"", ""text"": ""Code quality (beta): EXCELLENT""}, ""issues"": [{""code"": ""WPS432"", ""text"": ""Found magic number: 2"", ""line"": ""x = 2"", ""line_number"": 2, ""column_number"": 5, ""category"": ""BEST_PRACTICES"", ""difficulty"": ""MEDIUM"", ""influence_on_penalty"": 0}]}","{""quality"": {""code"": ""EXCELLENT"", ""text"": ""Code quality (beta): EXCELLENT""}, ""issues"": [{""code"": ""WPS432"", ""text"": ""Found magic number: 2"", ""line"": ""x = 2"", ""line_number"": 2, ""column_number"": 5, ""category"": ""BEST_PRACTICES"", ""difficulty"": ""MEDIUM"", ""influence_on_penalty"": 0}]}"
